Initial setting to incorporate region and offset domains into another domain
