+++ src/crab/abstract_domain.cpp
@@ -1,6 +1,7 @@
 #include "abstract_domain.hpp"
 #include "ebpf_domain.hpp"
 #include "type_domain.hpp"
+#include "region_domain.hpp"
 #include "offset_domain.hpp"
 
 template <typename Domain>
@@ -294,3 +295,4 @@ std::ostream& operator<<(std::ostream& o, const abstract_domain_t& dom) {
 template abstract_domain_t::abstract_domain_t(ebpf_domain_t);
 template abstract_domain_t::abstract_domain_t(type_domain_t);
 template abstract_domain_t::abstract_domain_t(offset_domain_t);
+template abstract_domain_t::abstract_domain_t(region_domain_t);
